Move visitor into its own file
